requirements.txt
# ============================================================
# Next-Bet-AI — Requirements
# Deep Learning Football Prediction Pipeline
# ============================================================

# --- Base de données (PostgreSQL) ---
sqlalchemy>=2.0
psycopg2-binary
alembic

# --- Manipulation de données ---
watchfiles==0.21.0
setuptools==75.8.2
feedparser==6.0.11
pandas
numpy

# --- API Backend ---
fastapi
uvicorn[standard]
pydantic>=2.0
stripe

# --- Scraping & HTTP ---
httpx
requests
understat
aiohttp>=3.13.0

# --- Deep Learning ---
torch
torchvision
transformers>=4.40       # BERT multilingue pour l'analyse de sentiment des news
sentencepiece            # tokenizer pour les modèles multilingues

# --- Machine Learning ---
xgboost>=2.0
lightgbm>=4.0

# --- Utilitaires ---
python-dotenv
scikit-learn

# --- Dev & Test ---
pytest
black